Bug fixes + Cleaned up the code and added comments
- Fixed a bug where the server returned an error if a user tried to login with the wrong email instead of returning a message to front-end.

+++ routes/articles.js
@@ -5,7 +5,6 @@ const authController = require('../controllers/auth');
 const marked = require('marked')
 const slugify = require('slugify')
 
-
 router.get('/blog-posts', authController.isLoggedIn, async (req, res) => {
     const articles = await Article.find().sort({ createdAt: 'desc'})
     res.render('articles/blog-posts.ejs', { 
@@ -13,18 +12,36 @@ router.get('/blog-posts', authController.isLoggedIn, async (req, res) => {
         user: req.user,
         message: ''
      })
-
-    // req.article = new Article()
 })
 
-router.get('/new', authController.isLoggedIn, (req, res) => {
-    res.render('articles/new.ejs', { article: new Article(), user: req.user })
-
+router.get('/new', authController.isLoggedIn, async (req, res) => {
+    const articles = await Article.find().sort({ createdAt: 'desc'})
+    if (req.user) {
+        res.render('articles/new.ejs', { article: new Article(), user: req.user })
+    } 
+    else {
+        res.render('articles/blog-posts.ejs', {
+            articles: articles,
+            user: null,
+            message: 'You need to be logged in to create a new post'
+        })
+    }
 })
 
 router.get('/edit/:id', authController.isLoggedIn, async (req, res) => {
+    const articles = await Article.find().sort({ createdAt: 'desc'})
     const article = await Article.findById(req.params.id)
-    res.render('articles/edit.ejs', { article: article, user: req.user })
+
+    if (req.user) {
+        res.render('articles/edit.ejs', { article: article, user: req.user })
+    }
+    else {
+        res.render('articles/blog-posts.ejs', {
+            articles: articles,
+            user: null,
+            message: 'You need to be logged in to edit a post'
+        })
+    }
 })
 
 router.put('/:id', async (req, res, next) => {
@@ -33,12 +50,18 @@ router.put('/:id', async (req, res, next) => {
 }, saveArticleAndRedirect('edit.ejs'))
 
 router.delete('/:id', authController.isLoggedIn, async (req, res) => {
-    await Article.findByIdAndDelete(req.params.id)
-    // res.redirect('/articles/blog-posts.ejs')
     const articles = await Article.find().sort({ createdAt: 'desc'})
-    res.render('articles/blog-posts.ejs', {articles: articles, user: req.user, message: ''})
-
-    // res.send('delete')
+    if (req.user) {
+        await Article.findByIdAndDelete(req.params.id)
+        res.redirect('/articles/blog-posts')
+    }
+    else {
+        res.render('articles/blog-posts.ejs', {
+            articles: articles,
+            user: null,
+            message: 'You need to be logged in to delete a post'
+        })
+    }
 })
 
 router.get('/:slug', authController.isLoggedIn, async (req, res) => {
@@ -48,30 +71,16 @@ router.get('/:slug', authController.isLoggedIn, async (req, res) => {
     else res.render('articles/show.ejs', { article: article, user: req.user })
 })
 
-
-
 router.post('/', (req, res, next) => {
-    // let article = new Article({
-    //     title: req.body.title,
-    //     description: req.body.description,
-    //     markdown: req.body.markdown
-    // })
-    // try {
-    //     article = await article.save()
-    //     res.redirect(`/articles/${article.slug}`)
-    // }
-    // catch (e) {
-    //     console.log(e)
-    //     res.render('articles/new.ejs', { article: article })
-    // }
     req.article = new Article()
     next()
 }, saveArticleAndRedirect('new.ejs'))
 
-
+// Used when creating new article or editing existing article,
+// gathers all data from body and saves it to database
 function saveArticleAndRedirect(path) {
     return async (req, res) => {
-        let article =req.article
+        let article = req.article
         
         article.title = req.body.title
         article.description = req.body.description
